minor improvements (#13)
* add file download component

* improve size regexes + parse as floats

+++ src/lib/utils/parsers.ts
@@ -94,11 +94,11 @@ const rgba = z
 
 export const color = z.union([hex, hsl, hsla, named, rgb, rgba]);
 
-const pixel = z.custom<`${number}px`>((val) => regex.pixel.test(val as string)).transform((val) => ({ value: parseInt(val.replace('px', '')), unit: 'px' }));
+const pixel = z.custom<`${number}px`>((val) => regex.pixel.test(val as string)).transform((val) => ({ value: parseFloat(val.replace('px', '')), unit: 'px' }));
 
 type pixel = z.infer<typeof pixel>;
 
-const point = z.custom<`${number}pt`>((val) => regex.point.test(val as string)).transform((val) => ({ value: parseInt(val.replace('pt', '')), unit: 'pt' }));
+const point = z.custom<`${number}pt`>((val) => regex.point.test(val as string)).transform((val) => ({ value: parseFloat(val.replace('pt', '')), unit: 'pt' }));
 
 type point = z.infer<typeof point>;
